Refactor FretesSolicitados component and related files
- Simplified the FretesSolicitados component by removing unused state variables and functions.
- Introduced AssociarCaminhaoForm component to handle the association of trucks with routes.
- Created CaminhoesVinculadosList component to display linked trucks.
- Added DialogConfirm component for confirmation dialogs when closing requests.
- Updated utility functions for calculating costs and ensuring route base IDs.
- Enhanced type definitions for better clarity and maintainability.
- Improved overall code organization and readability.
