+++ frontend/src/components/caminhoes/FretesSolicitados.tsx
@@ -1,116 +1,38 @@
 "use client";
 import { useEffect, useState } from "react";
-import { associarCaminhaoRota, atualizarCaminhaoRota, atualizarSolicitacaoRota, getCaminhaoRota, getRotas, getRotasSolicitadas, rotaBase } from "@/services/useFretesService";
+
+// Importando serviços e
+import { associarCaminhaoRota, atualizarCaminhaoRota, atualizarSolicitacaoRota, getCaminhaoRota, getRotasSolicitadas } from "@/services/useFretesService";
 import { getCaminhoes } from "@/services/useCarminhoesService";
 import { getParametrosFrete } from "@/services/useParametrosFretesService";
-import { Card, CardHeader, CardTitle, CardDescription, CardContent } from "@/components/ui/card";
-import { Pencil, Plus, Trash } from "lucide-react";
 import clsx from "clsx";
+
+// UI Components
+import { Pencil } from "lucide-react";
+
+import { Card, CardHeader, CardTitle, CardDescription, CardContent } from "@/components/ui/card";
 import {
   Dialog,
   DialogContent,
   DialogDescription,
-  DialogFooter,
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "@/components/ui/alert-dialog"
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
+
+// Types
 import { Caminhao } from "@/types/caminhoes";
 import { ParametrosFrete } from "@/types/Parametros";
 import { CaminhaoRotaVinculo } from "@/types/fretes";
+import { SolicitacaoFrete } from "@/types/solicitacaoFrete"
 
-interface CalcularCustosParams {
-  parametrosGlobais: ParametrosFrete | null;
-  kmTotal: number;
-  freteCaminhao: number;
-  pedagioIda: number;
-  pedagioVolta: number;
-  rotaBaseId: number | string | undefined;
-  caminhaoId: number | string;
-  consumoCombustivel?: number;
-  peso?: number;
-  diasViagem?: number;
-  qtdePneus?: number;
-}
+// Componentes Formação
+import { AssociarCaminhaoForm } from "./AssociarCaminhaoForm";
+import { CaminhoesVinculadosList } from "./CaminhoesVinculadosList"
+import { DialogConfirm } from "./ConfirmacaoDialog";
 
-function calcularCustosPorCaminhao({
-  parametrosGlobais,
-  kmTotal,
-  pedagioIda,
-  pedagioVolta,
-  rotaBaseId,
-  caminhaoId,
-  consumoCombustivel,
-  peso,
-  diasViagem,
-  qtdePneus,
-}: CalcularCustosParams) {
-  if (!diasViagem) {
-    diasViagem = Math.ceil(kmTotal / 500);
-  }
-
-  const qtdeDiesel = kmTotal / (consumoCombustivel ?? 1);
-  const custoCombustivel = qtdeDiesel * (parametrosGlobais?.valor_diesel_s10_sem_icms ?? 0);
-  const calcArla = (qtdeDiesel / (1000 / 50));
-  const custoArla = calcArla * 3.90;
-  const custoPneus = (parametrosGlobais?.valor_desgaste_pneus ?? 0) * kmTotal * (qtdePneus || 1);
-  const custoMotorista = diasViagem * (parametrosGlobais?.valor_salario_motorista_dia ?? 0);
-  const refeicaoMotorista = diasViagem * (parametrosGlobais?.valor_refeicao_motorista_dia ?? 0);
-  const ajudaCustoMotorista = diasViagem * (parametrosGlobais?.valor_ajuda_custo_motorista ?? 0);
-  const chapaDescarga = (parametrosGlobais?.valor_chapa_descarga ?? 0);
-
-  const custoTotal =
-    (
-      pedagioIda +
-      pedagioVolta +
-      custoCombustivel +
-      custoPneus +
-      custoMotorista +
-      refeicaoMotorista +
-      ajudaCustoMotorista +
-      chapaDescarga +
-      custoArla) * (1 + (parametrosGlobais?.margem_lucro ?? 0) / 100);
-
-  const custoPorKg = peso ? custoTotal / peso : custoTotal / 1;
-  return {
-    rota_base_id: rotaBaseId,
-    caminhao_id: caminhaoId,
-    pedagio_ida: pedagioIda,
-    pedagio_volta: pedagioVolta,
-    custo_combustivel: custoCombustivel,
-    custo_total: custoTotal,
-    salario_motorista_rota: custoMotorista,
-    refeicao_motorista_rota: refeicaoMotorista,
-    ajuda_custo_motorista_rota: ajudaCustoMotorista,
-    chapa_descarga_rota: chapaDescarga,
-    desgaste_pneus_rota: custoPneus,
-    custo_por_kg: custoPorKg,
-  };
-}
-
-type SolicitacaoFrete = {
-  id?: string;
-  origem: string;
-  destino: string;
-  peso: number;
-  status: string;
-  solicitante_user: string;
-  km_total?: number;
-  dias_viagem?: number;
-  rotaBaseId?: number;
-};
+// Funções 
+import { calcularCustosPorCaminhao } from "@/utils/calculoFrete"
+import { garantirRotaBaseId } from "@/utils/obtemIdRotas";
 
 const statusBgClasses: Record<string, string> = {
   PENDENTE: "bg-yellow-50 border-yellow-400",
@@ -130,24 +52,25 @@ export default function RotasSolicitadasList() {
 
   const [parametrosGlobais, setParametrosGlobais] = useState<ParametrosFrete | null>(null);
   const [rotaSelecionada, setRotaSelecionada] = useState<SolicitacaoFrete | null>(null);
-  const [kmTotal, setKmTotal] = useState("");
-  const [diasViagem, setDiasViagem] = useState(""); // Adicionado para controlar o input Dias de Viagem
   const [caminhoesDisponiveis, setCaminhoesDisponiveis] = useState<Caminhao[]>([]);
-  const [caminhaoSelecionado, setCaminhaoSelecionado] = useState("");
-  const [pedagioIda, setPedagioIda] = useState("");
-  const [pedagioVolta, setPedagioVolta] = useState("");
-  const [caminhoesVinculados, setCaminhoesVinculados] = useState<
-    { id: string; modelo: string; pedagioIda: number; pedagioVolta: number; consumo_medio: number; pneus: number }[]
-  >([]);
   const [caminhoesJaVinculados, setCaminhoesJaVinculados] = useState<CaminhaoRotaVinculo[]>([]);
-  const [mostrandoFormularioCaminhao, setMostrandoFormularioCaminhao] = useState(false);
-
-  type CustosPorCaminhao = ReturnType<typeof calcularCustosPorCaminhao>;
-  const [custosPorCaminhao, setCustosPorCaminhao] = useState<CustosPorCaminhao[]>([]);
   const [rotaBaseId, setRotaBaseId] = useState<number | null>(null);
+  const [dadosParaVincular, setDadosParaVincular] = useState<{
+    kmTotal: string;
+    diasViagem: string;
+    caminhoes: {
+      id: string;
+      modelo: string;
+      pedagioIda: number;
+      pedagioVolta: number;
+      consumo_medio: number;
+      pneus: number;
+    }[];
+  } | null>(null);
 
-  const [idParaFechar, setIdParaFechar] = useState<number | null>(null);
+
   const [openDialogConfirm, setOpenDialogConfirm] = useState(false);
+
 
   useEffect(() => {
     async function carregar() {
@@ -200,117 +123,13 @@ export default function RotasSolicitadasList() {
     }
   }
 
-  function handleFecharSolicitacao(rota: SolicitacaoFrete) {
-      setIdParaFechar(rota.id !== undefined ? Number(rota.id) : null);
-      setOpenDialogConfirm(true);
-    }
-
-    async function confirmarFechamento() {
-      if (idParaFechar) {
-        await atualizarSolicitacaoRota(Number(idParaFechar));
-        console.log("Solicitação de rota atualizada:", idParaFechar);
-        setOpenDialogConfirm(false);
-        setIdParaFechar(null);
-        await carregarRotasSolicitas();
-      }
-    }
-  function adicionarCaminhao() {
-    if (!caminhaoSelecionado || !pedagioIda || !pedagioVolta) return;
-
-    const caminhaoInfo = caminhoesDisponiveis.find((c) => c.id === caminhaoSelecionado);
-    if (!caminhaoInfo) return;
-
-    setCaminhoesVinculados((prev) => [
-      ...prev,
-      {
-        id: caminhaoInfo.id,
-        modelo: caminhaoInfo.modelo,
-        pedagioIda: parseFloat(pedagioIda),
-        pedagioVolta: parseFloat(pedagioVolta),
-        consumo_medio: caminhaoInfo.consumo_medio_km_l,
-        pneus: caminhaoInfo.pneus || 0,
-      },
-    ]);
-
-    setCaminhaoSelecionado("");
-    setPedagioIda("");
-    setPedagioVolta("");
-    setMostrandoFormularioCaminhao(false);
-  }
-
   function removerCaminhao(id: string | number) {
-    setCaminhoesVinculados((prev) => prev.filter((c) => c.id !== id));
+    setCaminhoesJaVinculados((prev) => prev.filter((c) => c.id !== id));
   }
-
-  // Função para garantir o rota_base_id (busca ou cria se não existir)
-  async function garantirRotaBaseId(origem: string, destino: string, km: number, dias: number): Promise<number | null> {
-    console.log("Garantindo rota base ID...");
-    if (!origem || !destino || !km || !dias) {
-      console.error("Parâmetros insuficientes para garantir rota base ID");
-      return null;
-    }
-    const rotas = await getRotas();
-    const rota = rotas.find((r) => r.destino === destino && r.origem === origem);
-    if (rota) {
-      const id = rota.id !== undefined ? Number(rota.id) : null;
-      setRotaBaseId(id);
-      console.log("Rota já existe, usando ID:", id);
-      return id;
-    } else {
-      const novaRota = await rotaBase(origem, destino, km, dias);
-      const id = novaRota.id !== undefined ? Number(novaRota.id) : null;
-      setRotaBaseId(id);
-      console.log("Nova rota criada, usando ID:", id);
-      return id;
-    }
-  }
-
   // Cálculo dos custos em tempo real para exibir na tela
-  useEffect(() => {
-    async function calcularCustos() {
-      if (
-        parametrosGlobais &&
-        kmTotal &&
-        caminhoesVinculados.length > 0 &&
-        rotaSelecionada?.origem &&
-        rotaSelecionada?.destino &&
-        diasViagem
-      ) {
-        const km = parseFloat(kmTotal);
-        const dias = parseFloat(diasViagem);
-        // Busca ou cria a rota antes do cálculo só para garantir id correto
-        const rota_base_id = await garantirRotaBaseId(rotaSelecionada.origem, rotaSelecionada.destino, km, dias);
 
-        const custos = caminhoesVinculados.map((caminhao) =>
-          calcularCustosPorCaminhao({
-            parametrosGlobais,
-            kmTotal: km,
-            freteCaminhao: 0,
-            pedagioIda: caminhao.pedagioIda,
-            pedagioVolta: caminhao.pedagioVolta,
-            rotaBaseId: rota_base_id ?? undefined,
-            caminhaoId: caminhao.id,
-            diasViagem: dias,
-            peso: rotaSelecionada?.peso,
-            consumoCombustivel: caminhao.consumo_medio,
-            qtdePneus: caminhao.pneus || 0,
-          })
-        );
-        setCustosPorCaminhao(custos);
-      } else {
-        setCustosPorCaminhao([]);
-      }
-    }
-    calcularCustos();
-  }, [parametrosGlobais, kmTotal, caminhoesVinculados, rotaSelecionada, diasViagem]);
-
-  async function salvarVinculo() {
-    if (
-      !rotaSelecionada?.origem ||
-      !rotaSelecionada?.destino ||
-      !kmTotal ||
-      !diasViagem
-    ) {
+  async function salvarVinculo(kmTotal: string, diasViagem: string, caminhoes: { id: string; modelo: string; pedagioIda: number; pedagioVolta: number; consumo_medio: number; pneus: number }[]) {
+    if (!rotaSelecionada?.origem || !rotaSelecionada?.destino || !kmTotal || !diasViagem) {
       alert("Preencha todos os campos obrigatórios!");
       return;
     }
@@ -318,15 +137,13 @@ export default function RotasSolicitadasList() {
     const parametros = await getParametrosFrete();
     const km = parseFloat(kmTotal);
     const dias = parseFloat(diasViagem);
-
-    // Garante que a rota existe e pega o id correto
     const rota_base_id = await garantirRotaBaseId(rotaSelecionada.origem, rotaSelecionada.destino, km, dias);
     console.log("Rota base ID para salvar Caminhão: ", rota_base_id);
-    // Carregue vínculos do backend como array de objetos completos!
     const caminhoesJaVinculados = await getCaminhaoRota(Number(rota_base_id));
 
-    const custosPorCaminhao = caminhoesVinculados.map((caminhao) =>
-      calcularCustosPorCaminhao({
+    for (const caminhao of caminhoes) {
+      // Calcule os custos para cada caminhão recebido do filho
+      const custo = calcularCustosPorCaminhao({
         parametrosGlobais: parametros,
         kmTotal: km,
         freteCaminhao: 0,
@@ -338,21 +155,13 @@ export default function RotasSolicitadasList() {
         peso: rotaSelecionada.peso,
         consumoCombustivel: caminhao.consumo_medio,
         qtdePneus: caminhao.pneus || 0,
-      })
-    );
+      });
 
-    console.log("Custos por caminhão calculados:", custosPorCaminhao);
-    for (const custo of custosPorCaminhao) {
-      console.log("Caminhões já vinculados:", caminhoesJaVinculados);
-
-      // Procura se já existe vínculo para este caminhão
       const jaVinculado = caminhoesJaVinculados.find(
         (v: { caminhao_id: string | number }) => String(v.caminhao_id) === String(custo.caminhao_id)
       );
 
       if (!jaVinculado) {
-        // Não existe vínculo: criar
-        console.log("Associando novo caminhão à rota base:", custo);
         await associarCaminhaoRota(
           Number(custo.rota_base_id),
           Number(custo.caminhao_id),
@@ -367,7 +176,6 @@ export default function RotasSolicitadasList() {
           Number(custo.desgaste_pneus_rota)
         );
       } else {
-        // Já existe, comparar os campos relevantes
         const dadosIguais =
           jaVinculado.pedagio_ida === custo.pedagio_ida &&
           jaVinculado.pedagio_volta === custo.pedagio_volta &&
@@ -380,28 +188,21 @@ export default function RotasSolicitadasList() {
           jaVinculado.desgaste_pneus_rota === custo.desgaste_pneus_rota;
 
         if (!dadosIguais) {
-          await atualizarCaminhaoRota(Number(jaVinculado.rota_base_id), Number(jaVinculado.caminhao_id),
-            {
-              pedagio_ida: custo.pedagio_ida,
-              pedagio_volta: custo.pedagio_volta,
-              custo_combustivel: custo.custo_combustivel,
-              custo_total: custo.custo_total,
-              salario_motorista_rota: custo.salario_motorista_rota,
-              refeicao_motorista_rota: custo.refeicao_motorista_rota,
-              ajuda_custo_motorista_rota: custo.ajuda_custo_motorista_rota,
-              chapa_descarga_rota: custo.chapa_descarga_rota,
-              desgaste_pneus_rota: custo.desgaste_pneus_rota
-            });
+          await atualizarCaminhaoRota(Number(jaVinculado.rota_base_id), Number(jaVinculado.caminhao_id), {
+            pedagio_ida: custo.pedagio_ida,
+            pedagio_volta: custo.pedagio_volta,
+            custo_combustivel: custo.custo_combustivel,
+            custo_total: custo.custo_total,
+            salario_motorista_rota: custo.salario_motorista_rota,
+            refeicao_motorista_rota: custo.refeicao_motorista_rota,
+            ajuda_custo_motorista_rota: custo.ajuda_custo_motorista_rota,
+            chapa_descarga_rota: custo.chapa_descarga_rota,
+            desgaste_pneus_rota: custo.desgaste_pneus_rota
+          });
         }
-        // Se for igual, não faz nada!
       }
     }
-
-    
-    setKmTotal("");
-    setDiasViagem("");
-    setCaminhoesVinculados([]);
-    setCustosPorCaminhao([]);
+    atualizarSolicitacaoRota(rotaSelecionada.id ? Number(rotaSelecionada.id) : 0)
     carregarRotasSolicitas();
   }
 
@@ -425,12 +226,9 @@ export default function RotasSolicitadasList() {
                   {rota.origem} → {rota.destino}
                 </CardTitle>
                 <button onClick={async () => {
-                  setRotaSelecionada(rota); // salva a solicitação (id)
-                  setKmTotal(String(rota.km_total || ""));
-                  setDiasViagem(String(rota.dias_viagem || ""));
+                  setRotaSelecionada(rota);
 
                   if (rota.rotaBaseId) {
-                    // Só busca caminhões vinculados SE existe rotaBaseId!
                     setRotaBaseId(Number(rota.rotaBaseId));
                     const vinculados = await getCaminhaoRota(Number(rota.rotaBaseId));
                     setCaminhoesJaVinculados(vinculados);
@@ -438,9 +236,6 @@ export default function RotasSolicitadasList() {
                     setRotaBaseId(null);
                     setCaminhoesJaVinculados([]);
                   }
-                  setMostrandoFormularioCaminhao(false);
-                  setCaminhoesVinculados([]);
-                  setCustosPorCaminhao([]);
                 }}>
                   <Pencil />
                 </button>
@@ -473,141 +268,50 @@ export default function RotasSolicitadasList() {
             </DialogDescription>
           </DialogHeader>
 
-          <div className="max-h-[500px] overflow-y-auto space-y-4 p-2">
-            <div>
-              <label className="text-sm font-medium">KM Total da Rota</label>
-              <Input
-                type="number"
-                value={kmTotal}
-                onChange={(e) => setKmTotal(e.target.value)}
-                placeholder="Ex: 320"
-              />
-            </div>
-            <div>
-              <label className="text-sm font-medium">Dias De Viagem</label>
-              <Input
-                type="number"
-                value={diasViagem}
-                onChange={(e) => setDiasViagem(e.target.value)}
-                placeholder="Ex: 5"
-              />
-            </div>
+          <AssociarCaminhaoForm
+            rotaSelecionada={rotaSelecionada || undefined}
+            caminhoesDisponiveis={caminhoesDisponiveis}
+            parametrosGlobais={parametrosGlobais}
+            onSolicitarFechamento={(dados) => {
+              setDadosParaVincular(dados);
+              setOpenDialogConfirm(true);
+            }}
+            onVincular={async () => {
+            }}
+          />
 
-            <div>
-              <label className="text-sm font-medium">Origem</label>
-              <Input type="text" value={rotaSelecionada?.origem || ""} disabled />
-            </div>
-
-            <div>
-              <label className="text-sm font-medium">Destino</label>
-              <Input type="text" value={rotaSelecionada?.destino || ""} />
-            </div>
-
-            {caminhoesJaVinculados.length > 0 && (
-              <div>
-                <p className="font-semibold text-sm mb-2">Caminhões já vinculados a esta rota</p>
-                <ul className="space-y-1">
-                  {caminhoesJaVinculados.map((c) => (
-                    <li
-                      key={c.caminhao_id}
-                      className="flex justify-between items-center border p-2 rounded-md text-sm"
-                    >
-                      <span>
-                        {/* Ajuste os campos conforme o objeto retornado do backend */}
-                        Caminhão: {c.modelo ?? c.caminhao_id} | Ida: R$ {Number(c.pedagio_ida).toFixed(2)} | Volta: R$ {Number(c.pedagio_volta).toFixed(2)}
-                      </span>
-                      {/* Se quiser permitir remover o vínculo direto daqui, implemente a função removerCaminhaoVinculado */}
-                      <Button
-                        variant="destructive"
-                        size="icon"
-                        onClick={() => removerCaminhao(c.caminhao_id)}
-                      >
-                        <Trash size={14} />
-                      </Button>
-                    </li>
-                  ))}
-                </ul>
-              </div>
-            )}
-
-            {custosPorCaminhao.length > 0 && (
-              <div>
-                <p className="font-semibold text-sm mb-2">Custo por caminhão (estimativa)</p>
-                <ul className="space-y-1">
-                  {custosPorCaminhao.map((c) => (
-                    <li key={c.caminhao_id} className="border p-2 rounded-md text-xs">
-                      <div>
-                        <strong>Caminhão:</strong> {c.caminhao_id}
-                      </div>
-                      <div>
-                        <strong>Custo total:</strong> R$ {c.custo_total.toFixed(2)}
-                      </div>
-                    </li>
-                  ))}
-                </ul>
-              </div>
-            )}
-
-            {!mostrandoFormularioCaminhao ? (
-              <Button onClick={() => setMostrandoFormularioCaminhao(true)} size="sm">
-                <Plus className="mr-2 h-4 w-4" /> Associar Caminhão
-              </Button>
-            ) : (
-              <div className="border p-3 rounded-md space-y-3">
-                <Select value={caminhaoSelecionado} onValueChange={(v) => setCaminhaoSelecionado(v)}>
-                  <SelectTrigger>
-                    <SelectValue placeholder="Selecione um caminhão" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    {caminhoesDisponiveis.map((c) => (
-                      <SelectItem key={c.id} value={c.id}>
-                        {c.modelo}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-
-                <Input
-                  type="number"
-                  placeholder="Pedágio ida"
-                  value={pedagioIda}
-                  onChange={(e) => setPedagioIda(e.target.value)}
+          {caminhoesJaVinculados.length > 0 && (
+            <>
+              <p className="font-semibold text-sm mb-2 mt-4">Caminhões já vinculados (salvos)</p>
+              <ul className="space-y-1">
+                <CaminhoesVinculadosList
+                  caminhoes={caminhoesJaVinculados}
+                  onRemover={removerCaminhao}
                 />
-                <Input
-                  type="number"
-                  placeholder="Pedágio volta"
-                  value={pedagioVolta}
-                  onChange={(e) => setPedagioVolta(e.target.value)}
-                />
-
-                <Button onClick={adicionarCaminhao} size="sm">
-                  Adicionar
-                </Button>
-              </div>
-            )}
-          </div>
-
-          <DialogFooter>
-            <Button onClick={() => {salvarVinculo(); handleFecharSolicitacao(rotaSelecionada as SolicitacaoFrete);}}>Salvar</Button>
-          </DialogFooter>
+              </ul>
+            </>
+          )}
         </DialogContent>
+
       </Dialog>
-      <AlertDialog open={openDialogConfirm} onOpenChange={setOpenDialogConfirm}>
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Deseja Fechar a Solicitação de Rota?</AlertDialogTitle>
-            <AlertDialogDescription>
-              Essa ação não pode ser desfeita.
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel onClick={() => setOpenDialogConfirm(false)}>Cancelar</AlertDialogCancel>
-            <AlertDialogAction onClick={confirmarFechamento}>
-              Confirmar
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
+
+      <DialogConfirm
+        openDialogConfirm={openDialogConfirm}
+        setOpenDialogConfirm={(open: boolean) => {
+          setOpenDialogConfirm(open);
+          if (!open) {
+            setDadosParaVincular(null);
+          }
+        }}
+        confirmarFechamento={async () => {
+          if (dadosParaVincular) {
+            await salvarVinculo(dadosParaVincular.kmTotal, dadosParaVincular.diasViagem, dadosParaVincular.caminhoes);
+            setOpenDialogConfirm(false);
+            setDadosParaVincular(null);
+            setRotaSelecionada(null);
+          }
+        }}
+      />
     </>
   );
 }
